requirements.txt
# --- Core Python & Data Libraries ---
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4
Pillow==10.3.0

# --- Deep Learning Framework ---
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

# --- Model Evaluation & Visualization ---
matplotlib==3.9.0
seaborn==0.13.2
scikit-learn==1.5.0

# --- Web Framework (Flask App) ---
Flask==3.0.3
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4

# --- Computer Vision & Image Processing ---
opencv-python==4.10.0.84

# --- Utility & Deployment ---
gunicorn==22.0.0  # Optional – for production hosting (e.g., Render, Heroku)

